made build and link work again.
- fixed all Archive methods with incorrect signatures
- fixed character replacement calls for BStrings
- added #ifdef around function declaration in MyListView that was causing
link error
- added missing -lstdc++ in Makefile
- removed extra cpp file in Makefile that isn't part of the main project

Notes:
- crashes on first startup due to missing preferences. Works on
2nd or 3rd run.
- missing some .jpg files so a few error dialogs pop up

+++ source/BeCalcWin.cpp
@@ -14,6 +14,7 @@
 #include "ExComplex.h"
 #include "Menus.h"
 #include "MBitmap.h"
+#include "MyListView.h"
 #include "Printer.h"
 #include "Reals.h"
 #include "Statistics.h"
@@ -373,7 +374,7 @@ void BeCalcWin::InitWindow()
 			case wReserved2:
 			case wNone: break;
 			default: // this is a plot window
-				doOpen.AddInt32("id", floor(log(select)/log(2))-5);
+				doOpen.AddInt32("id", floor(log((double)select)/log(2))-5);
 				doOpen.what = ID_REOPEN_PLOT;
 				break;
 		}
